renamed data_classes to data_types
