+++ src/arc/__init__.py
@@ -1,4 +1,5 @@
 from arc.config import Config
+
 from arc.cli import CLI
 from arc.utility import Utility
